requirements.txt
attrs==19.3.0
backcall==0.2.0
blis==0.4.1
brotlipy==0.7.0
catalogue @ file:///home/conda/feedstock_root/build_artifacts/catalogue_1589276142674/work
certifi==2020.6.20
cffi==1.11.5
chardet==3.0.4
click==7.1.2
cryptography==2.9.2
cymem @ file:///home/conda/feedstock_root/build_artifacts/cymem_1591801095798/work
datefinder==0.7.1
dateparser==0.7.6
decorator==4.4.2
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-2.3.0/en_core_web_sm-2.3.0.tar.gz
entrypoints==0.3
Flask==1.1.2
Flask-Cors==3.0.8
gevent==20.6.2
greenlet==0.4.16
gunicorn==20.0.4
idna==2.9
importlib-metadata @ file:///home/conda/feedstock_root/build_artifacts/importlib-metadata_1591451746006/work
ipykernel==5.3.0
ipython @ file:///tmp/build/80754af9/ipython_1592348391367/work
ipython-genutils==0.2.0
itsdangerous==1.1.0
jedi @ file:///tmp/build/80754af9/jedi_1592841914522/work
Jinja2==2.11.2
joblib==0.16.0
jsonschema==3.2.0
jupyter-client==6.1.3
jupyter-core==4.6.3
MarkupSafe==1.1.1
meinheld==1.0.2
murmurhash==1.0.2
mysql-connector==2.2.9
numpy @ file:///home/conda/feedstock_root/build_artifacts/numpy_1591485220230/work
pandas==1.0.4
parso==0.7.0
pendulum==2.1.1
pexpect==4.8.0
pickleshare==0.7.5
plac==1.1.3
preshed @ file:///home/conda/feedstock_root/build_artifacts/preshed_1591802139444/work
prompt-toolkit==3.0.5
ptyprocess==0.6.0
pycparser==2.20
Pygments==2.6.1
pyOpenSSL==19.1.0
pyrsistent==0.16.0
PySocks==1.7.1
python-dateutil==2.8.1
pytz==2020.1
pytzdata==2020.1
pyzmq==19.0.1
regex==2020.6.8
requests @ file:///home/conda/feedstock_root/build_artifacts/requests_1592425495151/work
scikit-learn==0.23.1
scipy==1.5.1
six==1.15.0
sklearn==0.0
spacy @ file:///home/conda/feedstock_root/build_artifacts/spacy_1592384903298/work
srsly @ file:///home/conda/feedstock_root/build_artifacts/srsly_1589222435996/work
thinc @ file:///home/conda/feedstock_root/build_artifacts/thinc_1590400402606/work
threadpoolctl==2.1.0
tornado==6.0.4
tqdm @ file:///home/conda/feedstock_root/build_artifacts/tqdm_1591181521996/work
traitlets==4.3.3
tzlocal==2.1
urllib3==1.25.9
wasabi @ file:///home/conda/feedstock_root/build_artifacts/wasabi_1592819210112/work
wcwidth==0.2.4
Werkzeug==1.0.1
zipp==3.1.0
zope.event==4.4
zope.interface==5.1.0
